Move decryption out of paasta local-run cmd
A follow-up PR will make paasta spark-run to use the helper function.
