+++ tests/cli/test_cmds_local_run.py
@@ -20,8 +20,6 @@ from pytest import raises
 from paasta_tools.adhoc_tools import AdhocJobConfig
 from paasta_tools.cli.cli import main
 from paasta_tools.cli.cmds.local_run import configure_and_run_docker_container
-from paasta_tools.cli.cmds.local_run import decrypt_secret_environment_for_service
-from paasta_tools.cli.cmds.local_run import decrypt_secret_environment_variables
 from paasta_tools.cli.cmds.local_run import docker_pull_image
 from paasta_tools.cli.cmds.local_run import format_command_for_type
 from paasta_tools.cli.cmds.local_run import get_container_id
@@ -37,7 +35,6 @@ from paasta_tools.cli.cmds.local_run import run_docker_container
 from paasta_tools.cli.cmds.local_run import run_healthcheck_on_container
 from paasta_tools.cli.cmds.local_run import simulate_healthcheck_on_service
 from paasta_tools.marathon_tools import MarathonServiceConfig
-from paasta_tools.secret_tools import SHARED_SECRET_SERVICE
 from paasta_tools.utils import InstanceConfig
 from paasta_tools.utils import NoConfigurationForServiceError
 from paasta_tools.utils import SystemPaastaConfig
@@ -1952,90 +1949,3 @@ def test_missing_volumes_skipped(mock_exists):
         )
         args, kwargs = mock_run_docker_container.call_args
         assert kwargs["volumes"] == []
-
-
-@mock.patch("paasta_tools.cli.cmds.local_run.is_secret_ref", autospec=True)
-@mock.patch("paasta_tools.cli.cmds.local_run.is_shared_secret", autospec=True)
-@mock.patch(
-    "paasta_tools.cli.cmds.local_run.decrypt_secret_environment_for_service",
-    autospec=True,
-)
-def test_decrypt_secret_environment_variables(
-    mock_decrypt_for_service, mock_is_shared_secret, mock_is_secret_ref
-):
-    mock_environment = {
-        "MY": "aaa",
-        "SECRET": "SECRET(123)",
-        "SECRET_SHARED": "SHARED_SECRET(abc)",
-    }
-    mock_is_secret_ref.side_effect = lambda val: "SECRET" in val
-    mock_is_shared_secret.side_effect = lambda val: "SHARED" in val
-    mock_decrypt_for_service.side_effect = [{"SECRET": "123"}, {"SECRET_SHARED": "abc"}]
-
-    ret = decrypt_secret_environment_variables(
-        secret_provider_name="vault",
-        environment=mock_environment,
-        soa_dir="/nail/blah",
-        service_name="universe",
-        cluster_name="mesosstage",
-        secret_provider_kwargs={"some": "config"},
-    )
-    assert ret == {"SECRET": "123", "SECRET_SHARED": "abc"}
-
-    expected_shared_args = dict(
-        secret_provider_name="vault",
-        soa_dir="/nail/blah",
-        cluster_name="mesosstage",
-        secret_provider_kwargs={"some": "config", "vault_num_uses": 2},
-    )
-    assert mock_decrypt_for_service.call_args_list == [
-        mock.call({"SECRET": "SECRET(123)"}, "universe", **expected_shared_args),
-        mock.call(
-            {"SECRET_SHARED": "SHARED_SECRET(abc)"},
-            SHARED_SECRET_SERVICE,
-            **expected_shared_args,
-        ),
-    ]
-
-
-@mock.patch(
-    "paasta_tools.cli.cmds.local_run.decrypt_secret_environment_for_service",
-    autospec=True,
-)
-def test_decrypt_secret_environment_variables_exception(mock_decrypt_for_service):
-    mock_decrypt_for_service.side_effect = KeyError
-    with raises(SystemExit):
-        decrypt_secret_environment_variables(
-            secret_provider_name="vault",
-            environment={},
-            soa_dir="/nail/blah",
-            service_name="universe",
-            cluster_name="mesosstage",
-            secret_provider_kwargs={"some": "config"},
-        )
-
-
-@mock.patch("paasta_tools.cli.cmds.local_run.get_secret_provider", autospec=True)
-def test_decrypt_secret_environment_for_service(mock_get_secret_provider):
-    mock_secret_env = {"SECRET": "SECRET(123)"}
-    mock_secret_provider = mock.Mock()
-    mock_get_secret_provider.return_value = mock_secret_provider
-    ret = decrypt_secret_environment_for_service(
-        secret_env_vars=mock_secret_env,
-        service_name="universe",
-        secret_provider_name="vault",
-        soa_dir="/nail/blah",
-        cluster_name="mesosstage",
-        secret_provider_kwargs={"some": "config"},
-    )
-    mock_get_secret_provider.assert_called_with(
-        secret_provider_name="vault",
-        soa_dir="/nail/blah",
-        service_name="universe",
-        cluster_names=["mesosstage"],
-        secret_provider_kwargs={"some": "config"},
-    )
-    mock_secret_provider.decrypt_environment.assert_called_with(
-        {"SECRET": "SECRET(123)"}
-    )
-    assert ret == mock_secret_provider.decrypt_environment.return_value
